output value of evidence.

+++ cyrecon.pxd
@@ -3,7 +3,7 @@ cdef extern from "PyFuncs.h":
   pass
 
 cdef extern from "proto.h":
-  int recon_run(int argc, char **argv)
+  double recon_run(int argc, char **argv)
 
 cdef extern from "allvars.h":
   pass
